Move state handlers types to custom file in `typings` dir
