+++ src/stateHandlers/ChoiceStateHandler.ts
@@ -1,10 +1,9 @@
 import type { JSONValue } from '../typings/JSONValue';
 import type { ChoiceState, ChoiceRuleWithoutNext } from '../typings/ChoiceState';
-import { BaseStateHandler, ExecutionResult } from './BaseStateHandler';
+import type { ChoiceStateHandlerOptions, ExecutionResult } from '../typings/StateHandlers';
+import { BaseStateHandler } from './BaseStateHandler';
 import { jsonPathQuery } from '../JsonPath';
 import wcmatch from 'wildcard-match';
-
-type ChoiceStateHandlerOptions = Record<string, unknown>;
 
 class ChoiceStateHandler extends BaseStateHandler<ChoiceState> {
   constructor(stateDefinition: ChoiceState) {
