+++ src/stateHandlers/FailStateHandler.ts
@@ -1,8 +1,7 @@
 import type { JSONValue } from '../typings/JSONValue';
 import type { FailState } from '../typings/FailState';
-import { BaseStateHandler, ExecutionResult } from './BaseStateHandler';
-
-type FailStateHandlerOptions = Record<string, unknown>;
+import type { ExecutionResult, FailStateHandlerOptions } from '../typings/StateHandlers';
+import { BaseStateHandler } from './BaseStateHandler';
 
 class FailStateHandler extends BaseStateHandler<FailState> {
   constructor(stateDefinition: FailState) {
