+++ src/stateHandlers/WaitStateHandler.ts
@@ -1,12 +1,9 @@
 import type { JSONValue } from '../typings/JSONValue';
 import type { WaitState } from '../typings/WaitState';
-import { BaseStateHandler, ExecutionResult } from './BaseStateHandler';
+import type { ExecutionResult, WaitStateHandlerOptions } from '../typings/StateHandlers';
+import { BaseStateHandler } from './BaseStateHandler';
 import { jsonPathQuery } from '../JsonPath';
 import { sleep } from '../util';
-
-type WaitStateHandlerOptions = {
-  waitTimeOverrideOption: number | undefined;
-};
 
 class WaitStateHandler extends BaseStateHandler<WaitState> {
   constructor(stateDefinition: WaitState) {
